make boto3 an optional dependency
