+++ rio_tiler/utils.py
@@ -8,7 +8,6 @@ from typing import Any, Dict, Generator, List, Optional, Sequence, Tuple, Union
 import numpy
 import rasterio
 from affine import Affine
-from boto3.session import Session as boto3_session
 from rasterio import windows
 from rasterio.crs import CRS
 from rasterio.dtypes import _gdal_typename
@@ -30,34 +29,6 @@ def _chunks(my_list: Sequence, chuck_size: int) -> Generator[Sequence, None, Non
     """Yield successive n-sized chunks from l."""
     for i in range(0, len(my_list), chuck_size):
         yield my_list[i : i + chuck_size]
-
-
-def aws_get_object(
-    bucket: str,
-    key: str,
-    request_pays: bool = False,
-    client: boto3_session.client = None,
-) -> bytes:
-    """AWS s3 get object content."""
-    if not client:
-        session = boto3_session()
-        # AWS_S3_ENDPOINT and AWS_HTTPS are GDAL config options of vsis3 driver
-        # https://gdal.org/user/virtual_file_systems.html#vsis3-aws-s3-files
-        endpoint_url = os.environ.get("AWS_S3_ENDPOINT", None)
-        if endpoint_url is not None:
-            use_https = os.environ.get("AWS_HTTPS", "YES")
-            if use_https.upper() in ["YES", "TRUE", "ON"]:
-                endpoint_url = "https://" + endpoint_url
-            else:
-                endpoint_url = "http://" + endpoint_url
-        client = session.client("s3", endpoint_url=endpoint_url)
-
-    params = {"Bucket": bucket, "Key": key}
-    if request_pays or os.environ.get("AWS_REQUEST_PAYER", "").lower() == "requester":
-        params["RequestPayer"] = "requester"
-
-    response = client.get_object(**params)
-    return response["Body"].read()
 
 
 def get_array_statistics(
